src/auth/requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.4
filelock==3.7.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
platformdirs==2.5.2
PyJWT==2.6.0
six==1.16.0
virtualenv==20.14.1
Werkzeug==2.2.2
